+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 9. when `FormeFieldController` or `FormeController ` is referenced  by others , `FormeFieldState`&`_FormeState` can be recycled after disposed
 10. remove `compareValue` on `FormeFieldState`
 11. remove `onFieldsRegistered` on `Forme`,use `onFieldsRegistered` & `onFieldsUnregistered` instead
+12. add `addVisitor` and `removeVisitor` on `FormeController`
 
 ## 3.5.2